// src/app/models/attendance.model.ts
export class Attendance {
  id?: any;
  semester?: string;
  date?: string;
  startTime?: string;
  endTime?: string;
  degreeCourse?: string;
  faculty?: string;
  mathBasic?: boolean;
  mathLow?: boolean;
  mathHigh?: boolean;
  programming?: boolean;
  physics?: boolean;
  chemistry?: boolean;
  organizational?: boolean;
  tutor?: any;
}